fix: Правки по результам ревью
