+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,20 +1,15 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.sevice.FilmService;
-import ru.yandex.practicum.filmorate.validator.EntityValidator;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 
 @RestController
@@ -45,39 +40,10 @@ public class FilmController {
         return service.putFilm(film);
     }
 
-    @PutMapping(value = "/films/{id}/like/{userId}")
-    public void addLike(@PathVariable(name = "id") @Positive long filmId, @PathVariable @Positive long userId) {
-        service.addLike(filmId, userId);
-    }
-
-    @DeleteMapping(value = "/films/{id}/like/{userId}")
-    public void removeLike(@PathVariable(name = "id") @Positive long filmId, @PathVariable @Positive long userId) {
-        service.removeLike(filmId, userId);
-    }
 
     @GetMapping(value = "/films/popular")
     public Set<Film> getPopular(@RequestParam(required = false, defaultValue = "10") @Positive int count) {
         return service.getOrderedSet(count);
-    }
-
-    @GetMapping(value = "/mpa")
-    public List<Film.Rating> getMpa(){
-        return service.getMpa();
-    }
-
-    @GetMapping(value = "/mpa/{id}")
-    public Film.Rating getMpa(@PathVariable @Positive long id){
-        return service.getMpa(id);
-    }
-
-    @GetMapping(value = "/genres")
-    public List<Film.Genre> getGenres(){
-        return service.getGenres();
-    }
-
-    @GetMapping(value = "/genres/{id}")
-    public Film.Genre getGenre(@PathVariable @Positive long id){
-        return service.getGenre(id);
     }
 
 }
