+++ src/main/java/ru/yandex/practicum/filmorate/model/Entity.java
@@ -2,9 +2,7 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.*;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.Optional;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,17 +11,9 @@ import java.util.stream.Stream;
 @EqualsAndHashCode
 public class Entity {
 
-    protected long id;
+    protected Optional<Long> id;
 
-    public static long findMissingId(List<Long> ids, long maxId) {
-        List<Long> seq = Stream.iterate(1L, n -> n + 1)
-                .limit(maxId)
-                .collect(Collectors.toList());
-        for (int i = 0; i < maxId - 1; i++) {
-            long xor = seq.get(i)^ids.get(i);
-            if (xor != 0) return seq.get(i);
-        }
-        return -1;
+    public Long getId() {
+        return id.orElse(null);
     }
-
 }
