+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -18,8 +18,8 @@ public class Film extends Entity implements Comparable<Film> {
 
     @Builder
     @ConstructorProperties({"id", "name", "description", "releaseDate", "duration", "rate", "genres", "mpa", "likes"})
-    public Film(long id, String name, @NonNull String description, @NonNull LocalDate releaseDate, @NonNull int duration, int rate, Set<Genre> genres, Rating mpa, Set<Long> likes) {
-        super(id);
+    public Film(long id, String name, @NonNull String description, @NonNull LocalDate releaseDate, @NonNull int duration, int rate, Set<Genre> genres, Mpa mpa, Set<Long> likes) {
+        super(Optional.of(id));
         this.name = name;
         this.description = description;
         this.releaseDate = releaseDate;
@@ -46,7 +46,7 @@ public class Film extends Entity implements Comparable<Film> {
 
     private int rate;
 
-    private Rating mpa;
+    private Mpa mpa;
 
     private Set<Genre> genres;
 
@@ -87,61 +87,6 @@ public class Film extends Entity implements Comparable<Film> {
         if (genres == null) {
             return Collections.emptySet();
         } else return new TreeSet<>(genres);
-    }
-
-    @Getter
-    public static class Genre implements Comparable {
-
-        @Positive
-        private final Optional<Long> id;
-        private Optional<String> name;
-
-        @ConstructorProperties("id")
-        public Genre(long id) {
-            this.id = Optional.of(id);
-        }
-
-        public Genre(long id, String name) {
-            this.id = Optional.of(id);
-            this.name = Optional.of(name);
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-            Genre genre = (Genre) o;
-            return Objects.equals(id, genre.id) && Objects.equals(name, genre.name);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(id, name);
-        }
-
-        @Override
-        public int compareTo(Object o) {
-            Genre genre = (Genre) o;
-            return (int) (this.id.orElse(0L) - genre.getId().orElse(0L));
-        }
-    }
-
-    @Getter
-    public static class Rating {
-
-        private final Optional<Long> id;
-        private Optional<String> name;
-
-        @ConstructorProperties("id")
-        public Rating(long id) {
-            this.id = Optional.of(id);
-        }
-
-
-        public Rating(long id, String name) {
-            this.id = Optional.of(id);
-            this.name = Optional.of(name);
-        }
     }
