+++ src/main/java/ru/yandex/practicum/filmorate/sevice/FilmService.java
@@ -1,14 +1,15 @@
 package ru.yandex.practicum.filmorate.sevice;
 
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.storage.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.Set;
+import java.util.TreeSet;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -17,9 +18,9 @@ public class FilmService {
 
     @Autowired
     @Qualifier("filmDbStorage")
-    final private FilmStorage filmStorage;
+    private final FilmStorage filmStorage;
 
-    public FilmService(@Qualifier("filmDbStorage") final FilmStorage filmStorage) {
+    public FilmService(@Autowired @Qualifier("filmDbStorage") final FilmStorage filmStorage) {
         this.filmStorage = filmStorage;
     }
 
@@ -39,24 +40,6 @@ public class FilmService {
         return filmStorage.putFilm(film);
     }
 
-    public void addLike(long filmId, long userId) {
-        Film film = filmStorage.getFilm(filmId);
-        if (film.getLikes().contains(userId)) {
-            log.warn("User with id " + userId + " has already liked this film");
-            return;
-        }
-        film.getLikes().add(userId);
-        filmStorage.putFilm(film);
-        log.info("User with id " + userId + " liked film by id " + filmId);
-
-    }
-
-    public void removeLike(long filmId, long userId) {
-        Film film = filmStorage.getFilm(filmId);
-        film.getLikes().remove(userId);
-        log.info("User with id " + userId + " removed liked for film by id " + filmId);
-    }
-
     public Set<Film> getOrderedSet(int range) {
         return new TreeSet<>(filmStorage.getFilms())
                 .stream()
@@ -64,19 +47,5 @@ public class FilmService {
                 .collect(Collectors.toSet());
     }
 
-    public List<Film.Rating> getMpa(){
-        return filmStorage.getMpas();
-    }
 
-    public Film.Rating getMpa(long id){
-        return filmStorage.getMpa(id);
-    }
-
-    public List<Film.Genre> getGenres(){
-        return filmStorage.getGenres();
-    }
-
-    public Film.Genre getGenre(long id){
-        return filmStorage.getGenre(id);
-    }
 }
