+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -1,9 +1,8 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.storage.film;
 
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.Collection;
-import java.util.List;
 
 public interface FilmStorage {
 
@@ -15,11 +14,4 @@ public interface FilmStorage {
 
     Film putFilm(Film film);
 
-    Film.Rating getMpa(long id);
-
-    List<Film.Rating> getMpas();
-
-    List<Film.Genre> getGenres();
-
-    Film.Genre getGenre(long id);
 }
