+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaStorage.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.filmorate.storage.mpa;
+
+import ru.yandex.practicum.filmorate.model.Mpa;
+
+import java.util.List;
+
+public interface MpaStorage {
+
+    Mpa getMpa(long id);
+
+    List<Mpa> getMpas();
+
+
+}
