+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserDbStorage.java
@@ -1,35 +1,37 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.storage.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.EntityAlreadyExistException;
 import ru.yandex.practicum.filmorate.exception.EntityIsNotFoundException;
-import ru.yandex.practicum.filmorate.model.Entity;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.storage.friend.FriendStorage;
 import ru.yandex.practicum.filmorate.validator.EntityValidator;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
 
 @Slf4j
-@Component("userDbStorage")
+@Component
 public class UserDbStorage implements UserStorage {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
+    @Autowired @Qualifier("friendDbStorage")
+    FriendStorage friendStorage;
+
     private long userId;
 
-    public long getNextId() {
-        SqlRowSet userIds = jdbcTemplate.queryForRowSet("select id from users");
-        List<Long> idList = new ArrayList<>();
-        while(userIds.next()){
-            idList.add(userIds.getLong("id"));
-        }
-        long id = Entity.findMissingId(idList, userId);
-        return id == -1 ? ++userId : id;
+    public UserDbStorage(@Autowired JdbcTemplate jdbcTemplate, @Autowired @Qualifier("friendDbStorage") FriendStorage friendStorage) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.friendStorage = friendStorage;
     }
 
     @Override
@@ -48,36 +50,10 @@ public class UserDbStorage implements UserStorage {
                             .build()
             );
         }
-        getFriends(users);
+        friendStorage.getFriends(users);
         return users.values();
     }
 
-    private void getFriends(Map<Long, User> users) {
-        SqlRowSet friendshipRows = jdbcTemplate.queryForRowSet("select user_id, friend_id, is_accepted from friendship");
-        while (friendshipRows.next()) {
-            long id = friendshipRows.getLong("user_id");
-            if (users.containsKey(id)) {
-                users.get(id)
-                        .getFriends()
-                        .put(
-                                friendshipRows.getLong("friend_id"),
-                                friendshipRows.getBoolean("is_accepted")
-                        );
-            }
-        }
-    }
-
-    private void getFriends(User user) {
-        SqlRowSet friendshipRows = jdbcTemplate.queryForRowSet("select user_id, friend_id, is_accepted from friendship where user_id = ?", user.getId());
-        while (friendshipRows.next()) {
-            long id = friendshipRows.getLong("user_id");
-            user.getFriends()
-                    .put(
-                            friendshipRows.getLong("friend_id"),
-                            friendshipRows.getBoolean("is_accepted")
-                    );
-        }
-    }
 
     @Override
     public User getUser(long id) {
@@ -90,7 +66,7 @@ public class UserDbStorage implements UserStorage {
                     .name(userRows.getString("name"))
                     .birthday(userRows.getDate("birthday").toLocalDate())
                     .build();
-            getFriends(user);
+            friendStorage.getFriends(user);
             return user;
         }
         log.error("User with id: " + id + " does not exist.");
@@ -104,36 +80,16 @@ public class UserDbStorage implements UserStorage {
             throw new EntityAlreadyExistException("User with id: " + user.getId() + " already exists.");
         }
         if (!EntityValidator.isUserNameValid(user)) user.setName(user.getLogin());
-        user.setId(getNextId());
         jdbcTemplate.update(
-                "INSERT INTO users (id, email, login, name, birthday) VALUES (?, ?, ?, ?, ?)",
-                user.getId(),
+                "INSERT INTO users (email, login, name, birthday) VALUES (?, ?, ?, ?)",
                 user.getEmail(),
                 user.getLogin(),
                 user.getName(),
                 user.getBirthday()
         );
-        addFriends(user);
-        return user;
-    }
-
-    private void addFriends(User user) {
-        if (!EntityValidator.isUserNameValid(user)) user.setName(user.getLogin());
-        for (Map.Entry<Long, Boolean> f : user.getFriends().entrySet()) {
-            jdbcTemplate.update(
-                    "INSERT INTO friendship (user_id, friend_id, is_accepted) values (?, ?, ?)",
-                    user.getId(),
-                    f.getKey(),
-                    f.getValue()
-            );
-        }
-    }
-
-    private void removeFriends(User user) {
-        jdbcTemplate.update(
-                "DELETE FROM friendship WHERE user_id = ?",
-                user.getId()
-        );
+        user.setId(Optional.of(getLastId()));
+        friendStorage.addFriends(user);
+        return getUser(getLastId());
     }
 
     @Override
@@ -151,16 +107,24 @@ public class UserDbStorage implements UserStorage {
                 user.getBirthday(),
                 user.getId()
         );
-        removeFriends(user);
-        addFriends(user);
+        friendStorage.removeFriends(user);
+        friendStorage.addFriends(user);
         return user;
     }
 
-    private boolean checkForUserExistence(User user){
+    private boolean checkForUserExistence(User user) {
         SqlRowSet row = jdbcTemplate.queryForRowSet("select id from users where id = ?", user.getId());
-        if (row.next()){
+        if (row.next()) {
             return row.getLong("id") == user.getId();
         }
         return false;
+    }
+
+    private long getLastId(){
+        SqlRowSet row = jdbcTemplate.queryForRowSet("select max(id) as id from users");
+        if (row.next()) {
+            return row.getLong("id");
+        }
+        return -1;
     }
 }
