+++ src/main/java/ru/yandex/practicum/filmorate/validator/EntityValidator.java
@@ -43,7 +43,7 @@ public class EntityValidator {
     }
 
     public static void isMpaValid(Film film) {
-        if (film.getMpa().getId().isEmpty()) {
+        if (film.getMpa().getId() <= 0) {
             EntityIsNotValidException e = new EntityIsNotValidException("MPA is invalid.");
             log.error("MPA is invalid.. Film's id:" + film.getId());
             throw e;
