+++ src/main/resources/data.sql
@@ -1,13 +1,13 @@
-INSERT INTO genres (id, genre) values (1, 'Комедия');
-INSERT INTO genres (id, genre) values (2, 'Драма');
-INSERT INTO genres (id, genre) values (3, 'Мультфильм');
-INSERT INTO genres (id, genre) values (4, 'Триллер');
-INSERT INTO genres (id, genre) values (5, 'Документальный');
-INSERT INTO genres (id, genre) values (6, 'Экшн');
+INSERT INTO genres (genre) values ('Комедия');
+INSERT INTO genres (genre) values ('Драма');
+INSERT INTO genres (genre) values ('Мультфильм');
+INSERT INTO genres (genre) values ('Триллер');
+INSERT INTO genres (genre) values ('Документальный');
+INSERT INTO genres (genre) values ('Экшн');
 
-INSERT INTO ratings (id, rating) values (1, 'G');
-INSERT INTO ratings (id, rating) values (2, 'PG');
-INSERT INTO ratings (id, rating) values (3, 'PG-13');
-INSERT INTO ratings (id, rating) values (4, 'R');
-INSERT INTO ratings (id, rating) values (5, 'NC-17');
+INSERT INTO ratings (rating) values ('G');
+INSERT INTO ratings (rating) values ('PG');
+INSERT INTO ratings (rating) values ('PG-13');
+INSERT INTO ratings (rating) values ('R');
+INSERT INTO ratings (rating) values ('NC-17');
