+++ src/main/resources/schema.sql
@@ -1,44 +1,46 @@
 DROP ALL OBJECTS;
 
 CREATE TABLE IF NOT EXISTS users(
-	id bigint PRIMARY KEY,
+	id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
 	email varchar(255),
 	login varchar(255),
 	name varchar(255),
 	birthday date
 );
 
+CREATE TABLE IF NOT EXISTS ratings(
+  id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+  rating varchar(255)
+);
+
 CREATE TABLE IF NOT EXISTS films(
-	id bigint PRIMARY KEY,
+	id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
 	name varchar(255),
 	description varchar(255),
 	release date,
 	duration integer,
-	rating_id bigint
+	rating_id bigint REFERENCES ratings (id) ON DELETE RESTRICT
 );
 
+CREATE TABLE IF NOT EXISTS genres(
+  id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+  genre varchar(255)
+);
+
+
 CREATE TABLE IF NOT EXISTS filmGenres(
-    film_id bigint,
-    genre_id bigint
+    film_id bigint REFERENCES films (id) ON DELETE CASCADE,
+    genre_id bigint REFERENCES genres (id) ON DELETE RESTRICT
 );
 
 CREATE TABLE IF NOT EXISTS friendship(
-	user_id bigint,
-	friend_id bigint,
+	user_id bigint REFERENCES users (id) ON DELETE CASCADE,
+	friend_id bigint REFERENCES users (id) ON DELETE CASCADE,
 	is_accepted boolean
 );
 
 CREATE TABLE IF NOT EXISTS likes(
-	film_id bigint ,
-	user_id bigint
+	film_id bigint REFERENCES films (id) ON DELETE CASCADE ,
+	user_id bigint REFERENCES users (id) ON DELETE CASCADE
 );
 
-CREATE TABLE IF NOT EXISTS genres(
-  id bigint,
-  genre varchar(255)
-);
-
-CREATE TABLE IF NOT EXISTS ratings(
-  id bigint,
-  rating varchar(255)
-);
